Make boxing explicit
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1503287 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/math3/stat/Frequency.java
@@ -546,7 +546,7 @@ public class Frequency implements Serializable {
         final Iterator<Map.Entry<Comparable<?>, Long>> iter = other.entrySetIterator();
         while (iter.hasNext()) {
             final Map.Entry<Comparable<?>, Long> entry = iter.next();
-            incrementValue(entry.getKey(), entry.getValue());
+            incrementValue(entry.getKey(), entry.getValue().longValue());
         }
     }
